requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
datetime==5.5
typing==3.7.4.3